mandando soma
